+++ back-end/src/errors/CannotJoinGameError.ts
@@ -1,0 +1,10 @@
+class CannotJoinGameError extends Error {
+  constructor(msg?: string) {
+    super(msg);
+
+    // Set the prototype explicitly.
+    Object.setPrototypeOf(this, CannotJoinGameError.prototype);
+  }
+}
+
+export default CannotJoinGameError;
